Update user scores on match update

+++ api/models/user.go
@@ -5,6 +5,6 @@ import "go.mongodb.org/mongo-driver/bson/primitive"
 type User struct {
 	Id        primitive.ObjectID `json:"_id" bson:"_id"`
 	Name      string             `json:"name" bson:"name"`
-	Score     int                `json:"score" bson:"score"`
+	Score     float64            `json:"score" bson:"score"`
 	IsDeleted bool               `json:"isDeleted" bson:"isDeleted"`
 }
